refactor(scopes): define scopes in their own files if necessary
