+++ src/types/drivinglicence.ts
@@ -1,4 +1,4 @@
-import { MyInfoField, StringValue, NumberValue, CodeAndDesc } from "./base"
+import { MyInfoField, StringValue, NumberValue, CodeAndDesc, MyInfoAttribute } from "./base"
 
 export type StartEndDate = {
   startdate: StringValue
@@ -29,3 +29,14 @@ export type DrivingLicenceCustomFields = {
 }
 
 export type MyInfoDrivingLicence = MyInfoField<DrivingLicenceCustomFields>
+export type DrivingLicenceScope =
+  | `${MyInfoAttribute.DrivingLicence}.${Exclude<
+      keyof DrivingLicenceCustomFields,
+      'suspension' | 'disqualification' | 'revocation' | 'pdl' | 'qdl'
+    >}`
+  | `${MyInfoAttribute.DrivingLicence}.${
+      | 'suspension'
+      | 'disqualification'
+      | 'revocation'}.${keyof StartEndDate}`
+  | `${MyInfoAttribute.DrivingLicence}.pdl.${keyof PDL}`
+  | `${MyInfoAttribute.DrivingLicence}.qdl.${keyof QDL}`
