+++ src/types/sponsoredchildrenrecords.ts
@@ -1,4 +1,4 @@
-import { MyInfoField, StringValue, CodeAndDesc } from "./base"
+import { MyInfoField, StringValue, CodeAndDesc, MyInfoAttribute } from "./base"
 
 export type SponsoredChildCustomFields = {
   nric: StringValue
@@ -23,3 +23,4 @@ export type SponsoredChildBelow21 = SponsoredChildCustomFields
 export type SponsoredChildAbove21 = Pick<SponsoredChildCustomFields, 'nric'>
 
 export type MyInfoSponsoredChildrenRecord = MyInfoField<SponsoredChildBelow21 | SponsoredChildAbove21>
+export type SponsoredChildrenRecordsScope = `${MyInfoAttribute.SponsoredChildrenRecords}.${keyof SponsoredChildCustomFields}`
